Major reorganization of how benchmarks are executed.
The goal of this reorganization is to have a predictable caching
behavior for both database caches and the file system buffer caches that
is relatively easy to reason about.

The benchmark operations are instantiated at the beginning of the run,
and all operations in a block of operation factories that do not involve
updates + the next block of update operations that belong to a single
factory are initialized together, followed by a sequential database scan
in order to pollute the cache contents. Basically, the benchmark runner
initializes as many operations as possible before a cache pollution scan
in the way that it preserves semantics of the original benchmark
implementation while minimizing the number of cache pollution scans.